Give listing its own module, and wrap the store where it is built
Both listing functions lived in `file_formats/list_datasets.rs`, which
made `browse_datasets` read as churn in a file that had no reason to
change, and put a cross-format concern under `file_formats`.

`beacon-functions/src/listing/` now holds the whole subject: `provider`
for the table both functions return, `list_datasets` and
`browse_datasets` for the two argument readers, `args` for the literal
parsing they share.

Also fixes where the sharded listing is applied. `BigPageList` only
wrapped stores built by `object_store_registry`, which serves ad-hoc
`s3://` URLs. The datasets store is not one of those: beacon-server
builds it and `register_object_stores` registers it verbatim, so it kept
the one-request-per-1000-objects walk and the sharding never reached the
data anyone actually queries. It is wrapped in `build_datasets_store`
now, beside the builder that makes it.
