+++ beacon-db/beacon-core/src/crawler/engine.rs
@@ -1,7 +1,7 @@
 //! The crawl engine: turn a [`CrawlerDefinition`] into registered external tables.
 //!
 //! Reuses Beacon's existing primitives end-to-end:
-//! - [`beacon_functions::file_formats::list_datasets::list_datasets`] for scan +
+//! - [`beacon_functions::listing::list_datasets`] for scan +
 //!   per-format classification,
 //! - [`ExternalTableDefinition::build_provider`] for schema inference + partition
 //!   validation (the same code path used when loading persisted tables),
@@ -19,7 +19,7 @@ use beacon_datafusion_ext::table_ext::{ExternalTable, ExternalTableDefinition, T
 use datafusion::prelude::SessionContext;
 use serde::{Deserialize, Serialize};
 
-use beacon_functions::file_formats::list_datasets::list_datasets;
+use beacon_functions::listing::list_datasets;
 
 use crate::statement_plan::{upgrade_session, SessionCell};
