+++ beacon-db/beacon-functions/src/lib.rs
@@ -13,6 +13,7 @@ use crate::{
 pub mod blue_cloud;
 pub mod file_formats;
 pub mod geo;
+pub mod listing;
 pub mod metadata;
 pub mod util;
